DWIM 110 - PyG -> otras tareas ajustes

+++ resources/views/backend/sales/gastos/index.blade.php
@@ -599,7 +599,7 @@ $(document).ready(function () {
                 return; 
               }
             }
-            total += parseInt($(this).data('import'));
+            total += parseFloat($(this).data('import'));
           }
         });
         if (all)   $('#totalMounth').text($('#totalMounth').data('orig'));
